+++ minsu3d/common_ops/src/get_iou/get_iou.cu
@@ -9,7 +9,7 @@ All Rights Reserved 2020.
 #include "get_iou.h"
 
 
-__global__ void get_iou_cuda_(int nInstance, int nProposal, int *proposals_idx, int *proposals_offset, int16_t *instance_labels, int *instance_pointnum, float *proposals_iou){
+__global__ void get_iou_cuda_(int nInstance, int nProposal, long *proposals_idx, int *proposals_offset, int16_t *instance_labels, int *instance_pointnum, float *proposals_iou){
     for(int proposal_id = blockIdx.x; proposal_id < nProposal; proposal_id += gridDim.x){
         int start = proposals_offset[proposal_id];
         int end = proposals_offset[proposal_id + 1];
@@ -18,7 +18,7 @@ __global__ void get_iou_cuda_(int nInstance, int nProposal, int *proposals_idx, 
             int instance_total = instance_pointnum[instance_id];
             int intersection = 0;
             for(int i = start; i < end; i++){
-                int idx = proposals_idx[i];
+                long idx = proposals_idx[i];
                 if(instance_labels[idx] == instance_id){
                     intersection += 1;
                 }
@@ -33,6 +33,6 @@ __global__ void get_iou_cuda_(int nInstance, int nProposal, int *proposals_idx, 
 //input: instance_labels (N), long, 0~total_nInst-1, -1
 //input: instance_pointnum (total_nInst), int
 //output: proposals_iou (nProposal, total_nInst), float
-void get_iou_cuda(int nInstance, int nProposal, int *proposals_idx, int *proposals_offset, int16_t *instance_labels, int *instance_pointnum, float *proposals_iou){
+void get_iou_cuda(int nInstance, int nProposal, long *proposals_idx, int *proposals_offset, int16_t *instance_labels, int *instance_pointnum, float *proposals_iou){
     get_iou_cuda_<<<std::min(nProposal, (int)32768), std::min(nInstance, (int)256)>>>(nInstance, nProposal, proposals_idx, proposals_offset, instance_labels, instance_pointnum, proposals_iou);
 }
